+++ lib/v2/screens/settings/edit_name/interactor/usecases/update_profile_use_case.dart
@@ -1,11 +1,11 @@
 import 'package:async/async.dart';
 import 'package:meta/meta.dart';
-import 'package:seeds/v2/datasource/remote/api/eos_repository.dart';
+import 'package:seeds/v2/datasource/remote/api/profile_repository.dart';
 
 export 'package:async/src/result/result.dart';
 
 class UpdateProfileUseCase {
-  final EosRepository _eosRepository = EosRepository();
+  final ProfileRepository _eosRepository = ProfileRepository();
 
   Future<Result> run({@required String name, String accountName, String privateKey, String nodeEndpoint}) {
     return _eosRepository.updateProfile(
